New file to install with Pacman
Hi. As promised here is the script that makes it work on my machine :)

This will work onto distros with the Pacman package manager (Arch Linux, Antergos, ArchBang, Chakra, Manjaro, Parabola and even more!). It is tested but may need some tuning.

I removed the update-icon-caches because it exists only for Ubuntu. The icon should still work to my knowledge (but will be a little longer to load) and if any other programs update the icon-cache it'll be updated too.